feat: add ability to write to index based on application context
Previously the logstash configuration wrote all logs to a single index, but in reality you probably
wouldn't do this. You'd want at the very least your applications to each have their own index. Now
the logstash output checks the value of the ApplicationContext field, and writes to a specific index
based on the value. This may lead to a situation where the logstash configuration gets very long,
one block for every app, which could be a problem down the road. I tried to use the
ApplicationContext as a variable so you could do this in a single block of output, but this didn't
work because index names are not allowed to have the '.' character, which our applications do. Maybe
we'll find a better way eventually, but this works fine for now.