feat: add ability to log file info in ecs format directly from serilog
Well, this one turned out to be pretty tricky indeed. After much effort this is the solution I
finally settled on, and it does indeed work. All you have to do is add a structured log element with
the @ sign and one of the ECS types, and it will magically appear in your logs in ECS format. It
feels super clunky to have to do all this so hopefully they'll come up with something better
eventually.