feat: add ability to load dashboards into kibana automatically
Previously, if you destroyed and re-created the system, everything you created in kibana would be
destroyed. Figured out how to export everything to json (ndjson specifically), and then import it
back in using an init container. This was quite tricky to figure out, need to write some docs on it.