fix: dont create an empty field when incorrectly logging strings using ecs structured log fields
Previously, if you accidentially logged a string to one of the custom ECS fields, it would throw an
exception. Got that fixed, but then it would create an empty field instead, also not great. Got that
fixed too. Now, if you try to log a simple string to one of those fields, it simply gets tossed.
There's probably a discussion to be had around whether silently throwing away input is the best
approach, but it's good enough for now.