feat: stop logging ecs json to the console
Previously, ECS JSON was printed to the console like any other log message, which I guess you might
wanna do sometimes but I thought it would be nice to show how to not do that if you were so
inclined, and figured out a way to do it. Gonna be an expert by the time I'm done.