feat: add ability to start system with security enabled
Previously, everything was working great with all the security turned off. Now that I want to add
some APM functionality, it requires security is turned on, which created a bunch of new problems I
had to figure out. The roles.yml file was the key, although I ended up having to give the anonymous
user way more permissions than I would prefer, this was the lowest amount I could find that would
work. But now you can login to kibana with the elastic account and all the healthchecks are working
again, so it's progress. Now I have to figure out how to create a fleet in docker compose.